Reorganize layers into general emulation layers module
