+++ external/fv3fit/fv3fit/emulation/layers/fields.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 from typing import Optional
 
-from ..layers import get_norm_class, get_denorm_class, IncrementStateLayer
+from .normalization import get_norm_class, get_denorm_class
 
 
 class FieldInput(tf.keras.layers.Layer):
@@ -90,6 +90,30 @@ class FieldOutput(tf.keras.layers.Layer):
         return tensor
 
 
+class IncrementStateLayer(tf.keras.layers.Layer):
+    """
+    Layer for incrementing states with a tendency tensor
+
+    Attributes:
+        dt_sec: timestep delta in seconds
+    """
+
+    def __init__(self, dt_sec: int, *args, dtype=tf.float32, **kwargs):
+
+        self.dt_sec = tf.constant(dt_sec, dtype=dtype)
+        super().__init__(*args, **kwargs)
+
+    def call(self, initial: tf.Tensor, tendency: tf.Tensor) -> tf.Tensor:
+        """
+        Increment state with tendency * timestep
+
+        args:
+            tensors: Input state field and corresponding tendency tensor to
+                increment by.
+        """
+        return initial + tendency * self.dt_sec
+
+
 class IncrementedFieldOutput(tf.keras.layers.Layer):
     """
     Add input tensor to an output tensor using timestep increment.
@@ -166,78 +190,3 @@ class CombineInputs(tf.keras.layers.Layer):
             ]
 
         return tf.concat(inputs, axis=self.combine_axis)
-
-
-class RNNBlock(tf.keras.layers.Layer):
-    """
-    RNN connected to an optional MLP for prediction
-
-    Combines multiple tensors along a new trailing dimension and
-    operates in the reverse on the feature dimension, i.e., from
-    the top of the atmosphere to the bottom for microphysics data.
-    """
-
-    def __init__(
-        self,
-        *args,
-        channels: int = 256,
-        dense_width: int = 256,
-        dense_depth: int = 1,
-        activation: str = "relu",
-        **kwargs,
-    ):
-        """
-        Args:
-            channels: width of RNN layer
-            dense_width: width of MLP layer, only used if dense_depth > 0
-            dense_depth: number of MLPlayers connected to RNN output, set to
-                0 to disable
-            activation: activation function to use for RNN and MLP
-        """
-        super().__init__(*args, **kwargs)
-
-        self.rnn = tf.keras.layers.SimpleRNN(
-            channels, activation=activation, go_backwards=True
-        )
-        self.dense = MLPBlock(width=dense_width, depth=dense_depth)
-
-    def call(self, input):
-
-        rnn_out = self.rnn(input)
-
-        output = self.dense(rnn_out)
-
-        return output
-
-
-class MLPBlock(tf.keras.layers.Layer):
-    """MLP layer for basic NN predictions"""
-
-    def __init__(
-        self,
-        *args,
-        width: int = 256,
-        depth: int = 2,
-        activation: str = "relu",
-        **kwargs,
-    ):
-        """
-        Args:
-            width: width of MLP layer, only used if dense_depth > 0
-            depth: number of MLPlayers, set to 0 to disable
-            activation: activation function to use for RNN and MLP
-        """
-        super().__init__(*args, **kwargs)
-
-        self.dense = [
-            tf.keras.layers.Dense(width, activation=activation) for i in range(depth)
-        ]
-
-    def call(self, input):
-
-        outputs = input
-
-        for i in range(len(self.dense)):
-            outputs = self.dense[i](outputs)
-
-        return outputs
