+++ external/fv3fit/tests/emulation/layers/test_fields.py
@@ -1,13 +1,10 @@
-import pytest
 import numpy as np
 import tensorflow as tf
 
-from fv3fit.emulation.microphysics.layers import (
+from fv3fit.emulation.layers.fields import (
     CombineInputs,
     FieldInput,
     FieldOutput,
-    MLPBlock,
-    RNNBlock,
     IncrementedFieldOutput,
 )
 
@@ -95,34 +92,3 @@ def test_CombineInputs_expand():
 
     assert result.shape == (20, 4, 3)
     np.testing.assert_array_equal(result[..., 2], tensor)
-
-
-def test_MLPBlock():
-
-    mlp = MLPBlock(width=256, depth=3)
-    assert len(mlp.dense) == 3
-
-    tensor = _get_tensor(20, 3)
-    result = mlp(tensor)
-
-    assert result.shape == (20, 256)
-
-
-def test_MLPBlock_no_dense_layers():
-    mlp = MLPBlock(width=256, depth=0)
-
-    tensor = _get_tensor(20, 10)
-    result = mlp(tensor)
-
-    assert result.shape == (20, 10)
-
-
-@pytest.mark.parametrize("depth,expected_shp", [(1, (20, 64)), (0, (20, 128))])
-def test_RNNBlock(depth, expected_shp):
-
-    rnn = RNNBlock(channels=128, dense_width=64, dense_depth=depth)
-
-    tensor = _get_tensor(20, 10)
-    combined = CombineInputs(-1, expand_axis=-1)((tensor, tensor))
-    result = rnn(combined)
-    assert result.shape == expected_shp
